# docs/swagger.yaml
definitions:
  service.Pack:
    properties:
      box:
        example: 50
        format: uint
        type: integer
      quantity:
        example: 3
        format: uint
        type: integer
    type: object
  service.PackRequest:
    properties:
      items:
        example: 543
        format: uint
        type: integer
    type: object
  service.PackResponse:
    properties:
      packs:
        items:
          $ref: '#/definitions/service.Pack'
        type: array
    type: object
  service.badRequestError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad request
        type: string
    type: object
  service.internalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal server error
        type: string
    type: object
  service.methodNotAllowedError:
    properties:
      code:
        example: 405
        type: integer
      message:
        example: Method not allowed
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: oleg.balunenko@gmail.com
    name: Oleg Balunenko
  description: This is a simple API for packing orders
  license:
    name: MIT
    url: https://opensource.org/license/mit
  termsOfService: http://swagger.io/terms/
  title: Order Packer API
  version: "1.0"
paths:
  /api/v1/pack:
    post:
      consumes:
      - application/json
      description: Calculates the number of packs needed to ship to a customer
      operationId: "orderpacker-pack\tpost"
      parameters:
      - description: Request data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.PackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with packs data
          schema:
            $ref: '#/definitions/service.PackResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/service.badRequestError'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/service.methodNotAllowedError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/service.internalServerError'
      summary: Get the number of packs needed to ship to a customer
      tags:
      - pack
schemes:
- http
swagger: "2.0"